% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clarkeGenesSigs}
\alias{clarkeGenesSigs}
\title{clarkeGenesSigs: Extract gene sets as matrix or list from Clarke Suppl file}
\usage{
clarkeGenesSigs(id = "Symbol", matrix = TRUE, convertNA = FALSE)
}
\arguments{
\item{id}{Character either with value "Symbol" or "Ensembl". Return signature as gene symbols or EnsEMBL gene identifers}

\item{matrix}{logical, default TRUE, return matrix, if FALSE return list}

\item{convertNA}{Logical. Default is FALSE Replace NA with 0}
}
\value{
matrix or list with gene sets
}
\description{
clarkeGenesSigs: Extract gene sets as matrix or list from Clarke Suppl file
}
\examples{
clarkeGenesSigs(id= "Symbol",matrix=TRUE)
clarkeGenesSigs(id= "Ensembl",matrix=FALSE)

}
