% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.clarkeGenes}
\alias{.clarkeGenes}
\title{.clarkeGenes: Create RData file from Clarke Genes Supplementary materials}
\usage{
.clarkeGenes()
}
\value{
Generates .rda file
}
\description{
.clarkeGenes: Create RData file from Clarke Genes Supplementary materials
}
\examples{
.clarkeGenes()
data(clarkeGenes)
}
