% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{excludeZeroSumRowCol}
\alias{excludeZeroSumRowCol}
\title{excludeZeroSumRowCol: Remove rows and columns where the rowSum or colSum is zero}
\usage{
excludeZeroSumRowCol(x)
}
\arguments{
\item{x}{a matrix  or data.frame}
}
\value{
matrix where rowSum(x) and colSum(x) >0
}
\description{
excludeZeroSumRowCol: Remove rows and columns where the rowSum or colSum is zero
}
\examples{
mm<-matrix(c(1:30, rep(0, 10)), ncol=4, nrow=10)
mm[4,]<-0
mm
excludeZeroSumRowCol(mm)

}
