% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapMouse2Human}
\alias{mapMouse2Human}
\title{mapMouse2Human}
\usage{
mapMouse2Human(MGI)
}
\arguments{
\item{MGI}{character vector of mouse gene symbols}
}
\value{
data.frame with columns c( MGI.symbol, HGNC.symbol,  Gene.stable.ID)
which are mouse gene symbols, human gene symbols and human EnsEMBL gene ids
}
\description{
mapMouse2Human
}
\examples{
mapMouse2Human(c("Cd4", "Cd8a", "Cd44"))

}
